Search module refactor
