+++ shared/src/types/search/index.ts
@@ -1,1 +1,3 @@
-export * from './search-result';
+export * from './output';
+export * from './input';
+export * from './search-event';
