+++ shared/src/types/search/input.ts
@@ -1,0 +1,4 @@
+export type SearchInput = {
+  query: string;
+  limit?: number;
+};
