+++ shared/src/types/search/output.ts
@@ -1,0 +1,4 @@
+export type SearchOutput<T> = {
+  item: T;
+  similarity: number;
+};
