+++ shared/src/types/search/search-event/index.ts
