+++ shared/src/types/search/search-event/output.ts
@@ -1,0 +1,6 @@
+import { Event } from '../../event';
+import { SearchOutput } from '../output';
+
+export type SearchEventOutput = SearchOutput<
+  Omit<Event, 'descriptionEmbedding'>
+>;
