--- shared/src/types/search/search-result.ts
@@ -1,6 +1,0 @@
-import { Event } from '../event';
-
-export type SearchResult = {
-  event: Omit<Event, 'descriptionEmbedding'>;
-  similarity: number;
-};
