+++ shared/src/validations/search/index.ts
@@ -1,1 +1,2 @@
 export * from './search-events';
+export * from './input';
