+++ shared/src/validations/search/search-events/output.ts
@@ -1,18 +1,18 @@
 import { z } from 'zod';
 import { eventSchema } from '../../event/event';
-import { SearchResult } from '../../../types';
+import { SearchEventOutput } from '../../../types';
 
-export const searchResultSchema = z
-  .object({
-    event: eventSchema.omit({ descriptionEmbedding: true }),
-    similarity: z.number().min(0).max(1),
-  })
-  .describe('SearchResult:') satisfies z.ZodType<SearchResult>;
+export type SearchEventsOutput = {
+  results: SearchEventOutput[];
+};
 
 export const searchEventsOutputSchema = z
   .object({
-    results: z.array(searchResultSchema),
+    results: z.array(
+      z.object({
+        item: eventSchema.omit({ descriptionEmbedding: true }),
+        similarity: z.number().min(0).max(1),
+      }),
+    ),
   })
-  .describe('SearchEventsOutput:');
-
-export type SearchEventsOutput = z.infer<typeof searchEventsOutputSchema>;
+  .describe('SearchEventsOutput:') satisfies z.ZodType<SearchEventsOutput>;
